+++ src/plugin/narrow_phase.rs
@@ -51,7 +51,7 @@ impl RapierContext {
     ///
     /// If this returns `None`, there is no contact between the two colliders.
     /// If this returns `Some`, then there may be a contact between the two colliders. Check the
-    /// result [`ContactPair::has_any_active_collider`] method to see if there is an actual contact.
+    /// result [`ContactPairView::has_any_active_contact`] method to see if there is an actual contact.
     pub fn contact_pair(&self, collider1: Entity, collider2: Entity) -> Option<ContactPairView> {
         let h1 = self.entity2collider.get(&collider1)?;
         let h2 = self.entity2collider.get(&collider2)?;
@@ -165,10 +165,12 @@ impl<'a> ContactManifoldView<'a> {
         self.raw.data.normal.into()
     }
 
+    /// The contacts that will be seen by the constraints solver for computing forces.
     pub fn num_solver_contacts(&self) -> usize {
         self.raw.data.solver_contacts.len()
     }
 
+    /// Gets the i-th solver contact.
     pub fn solver_contact(&self, i: usize) -> Option<SolverContactView> {
         self.raw
             .data